_manager_lib/requirements.txt
arrow==0.15.4
attrdict==2.0.1
attrs==19.3.0
aws-lambda-builders==0.5.0
aws-sam-cli==0.31.1
aws-sam-translator==1.15.1
awscli==1.16.283
binaryornot==0.4.4
blessed==1.16.1
boto3==1.10.19
botocore==1.13.19
certifi==2019.9.11
cfn-flip==1.2.2
cfn-lint==0.25.1
chardet==3.0.4
chevron==0.13.1
Click==7.0
colorama==0.4.1
cookiecutter==1.6.0
dateparser==0.7.2
docker==4.1.0
docutils==0.15.2
Flask==1.0.4
future==0.18.2
idna==2.8
importlib-metadata==0.23
importlib-resources==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.3
jinja2-time==0.2.0
jmespath==0.9.4
jsonpatch==1.24
jsonpointer==2.0
jsonschema==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
poyo==0.5.0
prompt-toolkit==1.0.14
pyasn1==0.4.8
Pygments==2.4.2
PyInquirer==1.0.3
pyrsistent==0.15.5
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.1.2
regex==2019.11.1
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
serverlessrepo==0.1.9
six==1.13.0
tzlocal==2.0.0
urllib3==1.25.7
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.16.0
whichcraft==0.6.1
zipp==0.6.0
